+++ webapp/src/filelist.tsx
@@ -5,9 +5,6 @@ import * as data from "./data";
 import * as sui from "./sui";
 import * as pkg from "./package";
 import * as core from "./core";
-import * as dialogs from "./dialogs";
-import * as workspace from "./workspace";
-import * as cloudsync from "./cloudsync";
 
 type ISettingsProps = pxt.editor.ISettingsProps;
 
@@ -274,60 +271,6 @@ namespace custom {
                 {!meta.numErrors ? null : <span className='ui label red'>{meta.numErrors}</span>}
             </div>
             {showFiles ? pxt.Util.concat(pkg.allEditorPkgs().map(p => this.filesWithHeader(p))) : undefined}
-        </div>;
-    }
-}
-
-interface GithubTreeItemState {
-    pushPulling?: boolean;
-}
-
-export class GithubTreeItem extends sui.UIElement<ISettingsProps, GithubTreeItemState> {
-    constructor(props: ISettingsProps) {
-        super(props);
-        this.state = {};
-        this.handleClick = this.handleClick.bind(this);
-        this.handleButtonKeydown = this.handleButtonKeydown.bind(this);
-    }
-
-    private handleButtonKeydown(e: React.KeyboardEvent<HTMLElement>) {
-        e.stopPropagation();
-    }
-
-    private handleClick(e: React.MouseEvent<HTMLElement>) {
-        e.stopPropagation();
-        const { githubId } = this.props.parent.state.header;
-        if (githubId) {
-            pxt.tickEvent("github.filelist.nav")
-            const gitf = pkg.mainEditorPkg().lookupFile("this/" + pxt.github.GIT_JSON);
-            this.props.parent.setSideFile(gitf);
-        }
-    }
-
-    renderCore() {
-        const header = this.props.parent.state.header;
-        if (!header) return <div />;
-
-        const { githubId } = header;
-        const ghid = pxt.github.parseRepoId(githubId);
-        if (!ghid) return <div />;
-
-        const targetTheme = pxt.appTarget.appTheme;
-        const mainPkg = pkg.mainEditorPkg()
-        const meta: pkg.PackageMeta = ghid ? this.getData("open-pkg-meta:" + mainPkg.getPkgId()) : undefined;
-
-        return <div role="tree" className={`ui tiny vertical ${targetTheme.invertedMenu ? `inverted` : ''} menu filemenu landscape only hidefullscreen`}>
-            <div
-                key="github-status"
-                className="item link"
-                onClick={this.handleClick}
-                tabIndex={0}
-                role="button"
-                onKeyDown={sui.fireClickOnEnter}>
-                {ghid.project && ghid.tag ? `${ghid.project}${ghid.tag == "master" ? "" : `#${ghid.tag}`}` : ghid.fullName}
-                <i className="github icon" />
-                {meta && meta.numFilesGitModified ? <i className="up arrow icon" /> : undefined}
-            </div>
         </div>;
     }
 }
